Make `ivis.Ivis` picklable
`Ivis` was made picklable by following a similar methodology to what is currently seen in `Ivis.save_model()` and `Ivis.load_model()`. The difference is that it saves the files inside the object itself by encoding it in `base85` by means of the built-in library `base64`. As such, instead of having four different files, they are all encoded and saved as attributes of the object, making it picklable and resulting in a single file. This also solves #101.

The only caveat is that, as of this moment, the saved models are not compiled, which means that an annoying warning is issued whenever a model is saved or loaded.